demo1-phoenix-rag/pyproject.toml
[tool.poetry]
name = "demo1-phoenix-rag"
version = "0.1.0"
description = "RAG over tax policy docs with Arize Phoenix observability"
authors = []
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.15"
langchain = "^1.2.15"
langchain-community = "^0.4.1"
langchain-chroma = "^1.1.0"
langchain-ollama = "^1.1.0"
chromadb = "^1.5.8"
arize-phoenix-otel = "^0.15.0"
arize-phoenix-evals = "^3.0.0"
openinference-instrumentation-langchain = "^0.1.62"
opentelemetry-sdk = "^1.31.1"
opentelemetry-exporter-otlp-proto-http = "^1.31.1"
pypdf = "^5.4.0"
httpx = "^0.28.1"
python-dotenv = "^1.1.0"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
